backend/requirements.txt
# === Web Framework ===
Flask==3.1.1
Flask-JWT-Extended==4.7.1
Flask-CORS==6.0.1

# === Base de données ===
mysql-connector-python==9.4.0
PyMySQL==1.1.1
SQLAlchemy==2.0.42
Flask-MySQLdb==2.0.0

# === Dotenv (config .env) ===
python-dotenv==1.1.1

# === Auth / Validation ===
pydantic==2.11.7
pydantic_core==2.33.2

# === Data / Analyse / Affichage ===
pandas==2.3.1
matplotlib==3.10.5
tabulate==0.9.0
numpy==2.2.6

# === PDF et texte arabe ===
fpdf2==2.7.4
arabic-reshaper==3.0.0
python-bidi==0.6.6

# === IA / LLM ===
openai==1.99.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
tiktoken==0.5.1

# === Logging / Debugging ===
colorlog==6.8.0